Moved configuration to better location, try a fix for spaces
