+++ src/github/posting/comment.ts
@@ -1,4 +1,4 @@
-import { githubConfig, octokit } from '../configuration';
+import { githubConfig, octokit } from '../../configuration';
 import Logger from '../../helper/logger';
 import { Analysis } from '../../helper/interfaces';
 import { createErrorSummary } from './summary';
